FIX : bugs in DMF

+++ recbox/model/general_recommender/dmf.py
@@ -44,7 +44,7 @@ class DMF(GeneralRecommender):
         self.item_layers_dim = config['item_layers_dim']
         # The dimensions of the last layer of users and items must be the same
         assert self.user_layers_dim[-1] == self.item_layers_dim[-1]
-        self.min_y_hat = torch.tensor([config['min_y_hat']]).to(self.device)
+        self.min_y_hat = torch.tensor([config['min_y_hat']]).double().to(self.device)
         self.inter_matrix_type = config['inter_matrix_type']
 
         # generate intermediate data
@@ -97,13 +97,17 @@ class DMF(GeneralRecommender):
         matrix_01.index_put_((row_indices, col_indices), self.history_user_value[item].flatten())
         item = self.item_linear(matrix_01)
 
-        user = self.user_fc_layers(user)
-        item = self.item_fc_layers(item)
+        user = self.user_fc_layers(user).double()
+        item = self.item_fc_layers(item).double()
 
-        user = F.normalize(user, p=2, dim=1)   # after normalize the vector, cosine distance reduced to dot product.
+        # after normalize the vector, cosine distance reduced to dot product.
+        #vector = torch.cosine_similarity(user, item).view(-1)
+        #user.double()
+        user = F.normalize(user, p=2, dim=1)
         item = F.normalize(item, p=2, dim=1)
         vector = torch.mul(user, item).sum(dim=1)
 
+        #norm_user_output = torch.sqrt((user**))
         vector = torch.max(vector, self.min_y_hat)  # restrict the result to [0, 1].
         return vector
 
@@ -171,11 +175,11 @@ class DMF(GeneralRecommender):
     def full_sort_predict(self, interaction):
         user = interaction[self.USER_ID]
         u_embedding = self.get_user_embedding(user)
-        u_embedding = self.user_fc_layers(u_embedding)
+        u_embedding = self.user_fc_layers(u_embedding).double()
         u_embedding = F.normalize(u_embedding, p=2, dim=1)
 
         if self.i_embedding is None:
-            self.i_embedding = self.get_item_embedding()
+            self.i_embedding = self.get_item_embedding().double()
             self.i_embedding = F.normalize(self.i_embedding, p=2, dim=1)
 
         cos_similarity = torch.mm(u_embedding, self.i_embedding.t())
